[ALPS04267344] Wlan: Fix Coverity issues
[Details]
Fix printf arg type mismatch.

MTK-Commit-Id: 8d034f49909cfe38c07aae3420a027d0512f6409

CR-Id: ALPS04267344
Feature: Wi-Fi Driver
(cherry picked from commit c73a1112c593377ccaaad4b653f95e86a9451221)

+++ os/linux/hif/common/hal_pdma.c
@@ -1252,7 +1252,7 @@ bool halWpdmaAllocTxRing(struct GLUE_INFO *prGlueInfo, uint32_t u4Num,
 	}
 
 	DBGLOG(HAL, TRACE, "TxDescRing[%p]: total %lu bytes allocated\n",
-	       prTxDesc->AllocVa, (uint64_t)prTxDesc->AllocSize);
+	       prTxDesc->AllocVa, prTxDesc->AllocSize);
 
 	/* Save PA & VA for further operation */
 	RingBasePa = prTxDesc->AllocPa;
@@ -1318,7 +1318,7 @@ bool halWpdmaAllocRxRing(struct GLUE_INFO *prGlueInfo, uint32_t u4Num,
 	}
 
 	DBGLOG(HAL, TRACE, "RxDescRing[%p]: total %lu bytes allocated\n",
-		prRxDesc->AllocVa, (uint64_t)prRxDesc->AllocSize);
+		prRxDesc->AllocVa, prRxDesc->AllocSize);
 
 	/* Initialize Rx Ring and associated buffer memory */
 	RingBasePa = prRxDesc->AllocPa;
